# code/backend/metiche/rest/server/list_member.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"log"
	"net/http"

	"go.einride.tech/aip/filtering"
	"go.einride.tech/aip/ordering"

	membermodule "github.com/mklfarha/metiche/backend/core/module/member"
	"github.com/mklfarha/metiche/backend/core/module/member/types"
	main_entity "github.com/mklfarha/metiche/backend/entity/member"
)

func (h *handlers) ListMember(w http.ResponseWriter, r *http.Request) {
	defaultPageSize := int32(10)
	maxPageSize := int32(100)

	lp, err := parseListParams(r, defaultPageSize, maxPageSize)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	offset, err := parsePageTokenOffset(lp.pageToken)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid page token: %w", err))
		return
	}

	declarations := memberDeclarations()
	filter, err := filtering.ParseFilter(lp, declarations)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid filter expression: %w", err))
		return
	}

	orderBy, err := ordering.ParseOrderBy(lp)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid order_by expression: %w", err))
		return
	}

	req := types.ListRequest{
		Offset:                offset,
		PageSize:              lp.GetPageSize(),
		Filter:                filter,
		FilteringDeclarations: declarations,
		EnumValues:            memberEnumValues(),
		OrderBy:               orderBy,
		IncludeFields:         lp.includeFields,
		ExcludeFields:         lp.excludeFields,
	}

	var result types.ListResponse
	if lp.skipCache {
		result, err = h.core.Member().List(r.Context(), req, membermodule.WithSkipCache())
	} else {
		result, err = h.core.Member().List(r.Context(), req)
	}
	if err != nil {
		// A filter or order_by the clause builder cannot express is the caller's
		// error, not the server's: statusForError turns it into a 400.
		writeProblem(w, statusForError(err), err)
		return
	}

	resp := ListEnvelope[main_entity.Member]{
		Items: result.Member,
	}
	if result.HasNextPage {
		resp.NextPageToken = encodePageToken(offset + int64(lp.GetPageSize()))
	}

	writeJSON(w, http.StatusOK, resp)
}

func memberDeclarations() *filtering.Declarations {
	declarations, err := filtering.NewDeclarations(
		filtering.DeclareStandardFunctions(),
		filtering.DeclareIdent("true", filtering.TypeBool),
		filtering.DeclareIdent("false", filtering.TypeBool),

		// member

		filtering.DeclareIdent("id", filtering.TypeString),

		filtering.DeclareIdent("team_uuid", filtering.TypeString),

		filtering.DeclareIdent("key", filtering.TypeString),

		filtering.DeclareIdent("display_name", filtering.TypeString),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in memberEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("role", filtering.TypeString),
		filtering.DeclareIdent("MEMBER_ROLE_INVALID", filtering.TypeString),
		filtering.DeclareIdent("MEMBER_ROLE_OWNER", filtering.TypeString),
		filtering.DeclareIdent("MEMBER_ROLE_MEMBER", filtering.TypeString),

		filtering.DeclareIdent("last_seen_at", filtering.TypeTimestamp),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in memberEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("status", filtering.TypeString),
		filtering.DeclareIdent("RECORD_STATUS_INVALID", filtering.TypeString),
		filtering.DeclareIdent("RECORD_STATUS_ACTIVE", filtering.TypeString),
		filtering.DeclareIdent("RECORD_STATUS_INACTIVE", filtering.TypeString),

		filtering.DeclareIdent("created_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("updated_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("account_uuid", filtering.TypeString),

		filtering.DeclareIdent("revoked_at", filtering.TypeTimestamp),
	)
	if err != nil {
		log.Printf("error creating declarations for member: %v", err)
	}
	return declarations
}

// memberEnumValues maps each enum filter field to every accepted
// spelling of its values and the number the column stores. REST generates no
// protobuf enum types (an app built with --api rest has no proto package at
// all), so this table — not the filter declarations — is what makes an enum
// filter resolvable here.
func memberEnumValues() map[string]map[string]int64 {
	return map[string]map[string]int64{

		"role": {
			"invalid":             0,
			"MEMBER_ROLE_INVALID": 0,
			"owner":               1,
			"MEMBER_ROLE_OWNER":   1,
			"member":              2,
			"MEMBER_ROLE_MEMBER":  2,
		},

		"status": {
			"invalid":                0,
			"RECORD_STATUS_INVALID":  0,
			"active":                 1,
			"RECORD_STATUS_ACTIVE":   1,
			"inactive":               2,
			"RECORD_STATUS_INACTIVE": 2,
		},
	}
}