# code/backend/metiche/rest/server/list_notification_channel.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"log"
	"net/http"

	"go.einride.tech/aip/filtering"
	"go.einride.tech/aip/ordering"

	notification_channelmodule "github.com/mklfarha/metiche/backend/core/module/notification_channel"
	"github.com/mklfarha/metiche/backend/core/module/notification_channel/types"
	main_entity "github.com/mklfarha/metiche/backend/entity/notification_channel"
)

func (h *handlers) ListNotificationChannel(w http.ResponseWriter, r *http.Request) {
	defaultPageSize := int32(10)
	maxPageSize := int32(100)

	lp, err := parseListParams(r, defaultPageSize, maxPageSize)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	offset, err := parsePageTokenOffset(lp.pageToken)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid page token: %w", err))
		return
	}

	declarations := notification_channelDeclarations()
	filter, err := filtering.ParseFilter(lp, declarations)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid filter expression: %w", err))
		return
	}

	orderBy, err := ordering.ParseOrderBy(lp)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid order_by expression: %w", err))
		return
	}

	req := types.ListRequest{
		Offset:                offset,
		PageSize:              lp.GetPageSize(),
		Filter:                filter,
		FilteringDeclarations: declarations,
		EnumValues:            notification_channelEnumValues(),
		OrderBy:               orderBy,
		IncludeFields:         lp.includeFields,
		ExcludeFields:         lp.excludeFields,
	}

	var result types.ListResponse
	if lp.skipCache {
		result, err = h.core.NotificationChannel().List(r.Context(), req, notification_channelmodule.WithSkipCache())
	} else {
		result, err = h.core.NotificationChannel().List(r.Context(), req)
	}
	if err != nil {
		// A filter or order_by the clause builder cannot express is the caller's
		// error, not the server's: statusForError turns it into a 400.
		writeProblem(w, statusForError(err), err)
		return
	}

	resp := ListEnvelope[main_entity.NotificationChannel]{
		Items: result.NotificationChannel,
	}
	if result.HasNextPage {
		resp.NextPageToken = encodePageToken(offset + int64(lp.GetPageSize()))
	}

	writeJSON(w, http.StatusOK, resp)
}

func notification_channelDeclarations() *filtering.Declarations {
	declarations, err := filtering.NewDeclarations(
		filtering.DeclareStandardFunctions(),
		filtering.DeclareIdent("true", filtering.TypeBool),
		filtering.DeclareIdent("false", filtering.TypeBool),

		// notification_channel

		filtering.DeclareIdent("id", filtering.TypeString),

		filtering.DeclareIdent("team_uuid", filtering.TypeString),

		filtering.DeclareIdent("project_uuid", filtering.TypeString),

		filtering.DeclareIdent("key", filtering.TypeString),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in notification_channelEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("kind", filtering.TypeString),
		filtering.DeclareIdent("NOTIFICATION_CHANNEL_KIND_INVALID", filtering.TypeString),
		filtering.DeclareIdent("NOTIFICATION_CHANNEL_KIND_SLACK", filtering.TypeString),
		filtering.DeclareIdent("NOTIFICATION_CHANNEL_KIND_DISCORD", filtering.TypeString),
		filtering.DeclareIdent("NOTIFICATION_CHANNEL_KIND_WEBHOOK", filtering.TypeString),

		filtering.DeclareIdent("label", filtering.TypeString),

		filtering.DeclareIdent("target_url", filtering.TypeString),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in notification_channelEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("min_severity", filtering.TypeString),
		filtering.DeclareIdent("CONFLICT_SEVERITY_INVALID", filtering.TypeString),
		filtering.DeclareIdent("CONFLICT_SEVERITY_LOW", filtering.TypeString),
		filtering.DeclareIdent("CONFLICT_SEVERITY_MEDIUM", filtering.TypeString),
		filtering.DeclareIdent("CONFLICT_SEVERITY_HIGH", filtering.TypeString),
		filtering.DeclareIdent("CONFLICT_SEVERITY_CRITICAL", filtering.TypeString),

		filtering.DeclareIdent("notify_human_requests", filtering.TypeBool),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in notification_channelEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("status", filtering.TypeString),
		filtering.DeclareIdent("RECORD_STATUS_INVALID", filtering.TypeString),
		filtering.DeclareIdent("RECORD_STATUS_ACTIVE", filtering.TypeString),
		filtering.DeclareIdent("RECORD_STATUS_INACTIVE", filtering.TypeString),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in notification_channelEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("delivery_status", filtering.TypeString),
		filtering.DeclareIdent("DELIVERY_STATUS_INVALID", filtering.TypeString),
		filtering.DeclareIdent("DELIVERY_STATUS_NEVER_SENT", filtering.TypeString),
		filtering.DeclareIdent("DELIVERY_STATUS_OK", filtering.TypeString),
		filtering.DeclareIdent("DELIVERY_STATUS_FAILING", filtering.TypeString),
		filtering.DeclareIdent("DELIVERY_STATUS_DISABLED_AFTER_FAILURES", filtering.TypeString),

		filtering.DeclareIdent("consecutive_failures", filtering.TypeInt),

		filtering.DeclareIdent("last_attempt_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("last_success_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("last_error", filtering.TypeString),

		filtering.DeclareIdent("created_by_member_uuid", filtering.TypeString),

		filtering.DeclareIdent("created_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("updated_at", filtering.TypeTimestamp),
	)
	if err != nil {
		log.Printf("error creating declarations for notification_channel: %v", err)
	}
	return declarations
}

// notification_channelEnumValues maps each enum filter field to every accepted
// spelling of its values and the number the column stores. REST generates no
// protobuf enum types (an app built with --api rest has no proto package at
// all), so this table — not the filter declarations — is what makes an enum
// filter resolvable here.
func notification_channelEnumValues() map[string]map[string]int64 {
	return map[string]map[string]int64{

		"kind": {
			"invalid":                           0,
			"NOTIFICATION_CHANNEL_KIND_INVALID": 0,
			"slack":                             1,
			"NOTIFICATION_CHANNEL_KIND_SLACK":   1,
			"discord":                           2,
			"NOTIFICATION_CHANNEL_KIND_DISCORD": 2,
			"webhook":                           3,
			"NOTIFICATION_CHANNEL_KIND_WEBHOOK": 3,
		},

		"min_severity": {
			"invalid":                    0,
			"CONFLICT_SEVERITY_INVALID":  0,
			"low":                        1,
			"CONFLICT_SEVERITY_LOW":      1,
			"medium":                     2,
			"CONFLICT_SEVERITY_MEDIUM":   2,
			"high":                       3,
			"CONFLICT_SEVERITY_HIGH":     3,
			"critical":                   4,
			"CONFLICT_SEVERITY_CRITICAL": 4,
		},

		"status": {
			"invalid":                0,
			"RECORD_STATUS_INVALID":  0,
			"active":                 1,
			"RECORD_STATUS_ACTIVE":   1,
			"inactive":               2,
			"RECORD_STATUS_INACTIVE": 2,
		},

		"delivery_status": {
			"invalid":                    0,
			"DELIVERY_STATUS_INVALID":    0,
			"never_sent":                 1,
			"DELIVERY_STATUS_NEVER_SENT": 1,
			"ok":                         2,
			"DELIVERY_STATUS_OK":         2,
			"failing":                    3,
			"DELIVERY_STATUS_FAILING":    3,
			"disabled_after_failures":    4,
			"DELIVERY_STATUS_DISABLED_AFTER_FAILURES": 4,
		},
	}
}