# code/backend/metiche/rest/server/openapi.yaml
...
openapi: 3.1.0
info:
    title: Metiche
    version: 1.0.0
servers:
    - url: http://localhost:8080/v1
paths:
...
    /members:
        get:
            tags:
                - Member
            summary: List members
            operationId: listMember
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of members
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MemberList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - Member
            summary: Create Member
            operationId: createMember
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Member'
                required: true
            responses:
                "201":
                    description: Created Member
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Member'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /notification-channels:
        get:
            tags:
                - NotificationChannel
            summary: List notification-channels
            operationId: listNotificationChannel
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of notification-channels
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotificationChannelList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - NotificationChannel
            summary: Create NotificationChannel
            operationId: createNotificationChannel
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NotificationChannel'
                required: true
            responses:
                "201":
                    description: Created NotificationChannel
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotificationChannel'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
components:
    schemas:
...
        ConflictSeverity:
            enum:
                - invalid
                - low
                - medium
                - high
                - critical
            type: string
...
        DeliveryStatus:
            enum:
                - invalid
                - never_sent
                - ok
                - failing
                - disabled_after_failures
            type: string
...
        Member:
            properties:
                account_uuid:
                    format: uuid
                    type: string
                created_at:
                    format: date-time
                    type: string
                display_name:
                    type: string
                id:
                    format: uuid
                    type: string
                key:
                    type: string
                last_seen_at:
                    format: date-time
                    nullable: true
                    type: string
                revoked_at:
                    format: date-time
                    nullable: true
                    type: string
                role:
                    $ref: '#/components/schemas/MemberRole'
                status:
                    $ref: '#/components/schemas/RecordStatus'
                team_uuid:
                    format: uuid
                    type: string
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - team_uuid
                - key
                - display_name
                - role
                - status
                - created_at
                - updated_at
                - account_uuid
            type: object
        MemberList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/Member'
                    type: array
                next_page_token:
                    type: string
            type: object
        MemberRole:
            enum:
                - invalid
                - owner
                - member
            type: string
        NotificationChannel:
            properties:
                consecutive_failures:
                    format: int64
                    type: integer
                created_at:
                    format: date-time
                    type: string
                created_by_member_uuid:
                    format: uuid
                    nullable: true
                    type: string
                delivery_status:
                    $ref: '#/components/schemas/DeliveryStatus'
                id:
                    format: uuid
                    type: string
                key:
                    type: string
                kind:
                    $ref: '#/components/schemas/NotificationChannelKind'
                label:
                    type: string
                last_attempt_at:
                    format: date-time
                    nullable: true
                    type: string
                last_error:
                    nullable: true
                    type: string
                last_success_at:
                    format: date-time
                    nullable: true
                    type: string
                min_severity:
                    $ref: '#/components/schemas/ConflictSeverity'
                notify_human_requests:
                    type: boolean
                project_uuid:
                    format: uuid
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/RecordStatus'
                target_url:
                    type: string
                team_uuid:
                    format: uuid
                    type: string
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - team_uuid
                - key
                - kind
                - label
                - target_url
                - min_severity
                - notify_human_requests
                - status
                - delivery_status
                - consecutive_failures
                - created_at
                - updated_at
            type: object
        NotificationChannelKind:
            enum:
                - invalid
                - slack
                - discord
                - webhook
            type: string
        NotificationChannelList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/NotificationChannel'
                    type: array
                next_page_token:
                    type: string
            type: object
...
        Problem:
            properties:
                detail:
                    type: string
                instance:
                    type: string
                status:
                    type: integer
                title:
                    type: string
                type:
                    type: string
            type: object
...
        RecordStatus:
            enum:
                - invalid
                - active
                - inactive
            type: string
...